pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pennywise-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "Django>=5.0",
  "djangorestframework>=3.15",
  "drf-spectacular>=0.27",
  "psycopg2-binary>=2.9",
  "Pillow>=10.0",
  "celery>=5.3",
  "redis>=5.0",
  "boto3>=1.34",
  "django-storages>=1.14",
  "pydantic>=2.7",
  "faiss-cpu>=1.8.0"
]
